requirements.txt
django
psycopg2
djangorestframework
django-filter
django-cors-headers
PyJWT
pytz